+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [2023.2.1] - 2023-08-15
+
 ### Added
 - GC and exception markers to the chart
 - Automatically attach charts to a starting process 
@@ -82,7 +84,8 @@
 - Actions to monitor and collect counters
 - Action to collect traces
 
-[Unreleased]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.2.0...HEAD
+[Unreleased]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.2.1...HEAD
+[2023.2.1]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.2.0...v2023.2.1
 [2023.2.0]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.1.2...v2023.2.0
 [2023.1.2]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.1.1...v2023.1.2
 [2023.1.1]: https://github.com/rafaelldi/diagnostics-client-plugin/compare/v2023.1.0...v2023.1.1